backend/requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0

# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7

# Machine Learning (через API)
openai>=1.12.0
httpx>=0.25.0
# Для работы с данными (если нужно)
numpy==1.24.4
pandas==2.1.3
pillow==10.1.0

# Database
asyncpg==0.29.0
aioredis==2.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
